docstrings and tests for merge_marc code used by import
